Cachine of volume histos and vwap - work in progress
